etats-persistence-donnees/orm/djangoex/pyproject.toml
[project]
name = "djangoex"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.1",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.29.0",
    "drf-spectacular-sidecar>=2026.1.1",
]
